refactor [labels 1/4]: Group social functions (with colors)
Side effect is that the social colors won't be part of the color pool anymore
