+++ src/components/study/social-sustainability/SummaryBlock.vue
@@ -12,7 +12,7 @@
 
 <script setup>
 import { computed } from 'vue';
-import { getSocialScoreColor } from '@utils/colors.js'
+import { getSocialScoreColor } from '@utils/social.js'
 const props = defineProps({
   title: String,
   anchor: Number,
